requirements.txt
pip==24.3.1 
beautifulsoup4==4.12.3

Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-Login==0.6.3

passlib==1.7.4

pyarrow==18.0.0
pyOpenSSL==23.2.0
pymongo==4.10.1
python-dotenv==1.0.1
python-magic==0.4.27
pytest==8.3.4
requests==2.32.3
pillow
scipy==1.12.0
seaborn==0.13.2

virtualenv==20.25.0
zipp==3.20.2


scikit-learn==1.5.2
nltk==3.9.1


pinecone
pinecone-client
pinecone-plugin-inference

langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.7
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.4
langsmith==0.2.6

